refactor: 이미지 파일 즉시 삭제, 배치 삭제 로직 구분에 따른 수정 - 불필요한 로직 삭제 및 로직 변경에 따른 리팩토링 작업
