+++ src/main/java/com/thekey/stylekeyserver/brand/controller/BrandAdminController.java
@@ -4,10 +4,12 @@ import com.thekey.stylekeyserver.brand.domain.Brand;
 import com.thekey.stylekeyserver.brand.dto.request.BrandRequest;
 import com.thekey.stylekeyserver.brand.dto.response.BrandResponse;
 import com.thekey.stylekeyserver.brand.service.BrandAdminService;
-import com.thekey.stylekeyserver.common.ApiResponse;
-import com.thekey.stylekeyserver.common.ErrorCode;
+import com.thekey.stylekeyserver.common.exception.ApiResponse;
+import com.thekey.stylekeyserver.common.exception.ErrorCode;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import java.io.UnsupportedEncodingException;
+import java.net.MalformedURLException;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -93,7 +95,8 @@ public class BrandAdminController {
 
     @DeleteMapping("/{id}")
     @Operation(summary = "Delete Brand", description = "브랜드 정보 삭제")
-    public ApiResponse<Void> deleteBrand(@PathVariable Long id) {
+    public ApiResponse<Void> deleteBrand(@PathVariable Long id)
+            throws MalformedURLException, UnsupportedEncodingException {
         brandAdminService.delete(id);
         return ApiResponse.success();
     }
