+++ src/main/java/com/thekey/stylekeyserver/brand/service/BrandAdminService.java
@@ -2,6 +2,8 @@ package com.thekey.stylekeyserver.brand.service;
 
 import com.thekey.stylekeyserver.brand.domain.Brand;
 import com.thekey.stylekeyserver.brand.dto.request.BrandRequest;
+import java.io.UnsupportedEncodingException;
+import java.net.MalformedURLException;
 import java.util.List;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -17,6 +19,6 @@ public interface BrandAdminService {
 
     Brand update(Long id, BrandRequest requestDto, MultipartFile imageFile) throws Exception;
 
-    void delete(Long id);
+    void delete(Long id) throws MalformedURLException, UnsupportedEncodingException;
 
 }
