+++ src/main/java/com/thekey/stylekeyserver/coordinatelook/service/CoordinateLookAdminService.java
@@ -5,6 +5,8 @@ import com.thekey.stylekeyserver.coordinatelook.dto.response.CoordinateLookRespo
 import com.thekey.stylekeyserver.coordinatelook.dto.request.CoordinateLookRequest;
 import com.thekey.stylekeyserver.item.dto.request.ItemRequest;
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.net.MalformedURLException;
 import java.util.List;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -26,7 +28,7 @@ public interface CoordinateLookAdminService {
     CoordinateLook updateItem(Long coordinateLookId, Long itemId, ItemRequest requestDto,
                               MultipartFile itemImageFile) throws IOException;
 
-    void delete(Long id);
+    void delete(Long id) throws MalformedURLException, UnsupportedEncodingException;
 
     void deleteItemFromCoordinateLook(Long coordinateLookId, Long itemId);
 }
